feat: fixed navigation handling with callback
